requirements/base.txt
django==5.1
django-environ==0.11.2
django-allauth==64.0.0
django-filter==24.3
django-bootstrap5==24.2
django-crispy-forms==2.3

fido2==1.1.3